Move kubernetes code into a backend class
