+++ tests/projects/test_projects.py
@@ -10,6 +10,7 @@ import mlflow
 
 from mlflow.entities import RunStatus, ViewType, SourceType
 from mlflow.exceptions import ExecutionException, MlflowException
+from mlflow.projects.backend.kubernetes import _parse_kubernetes_config
 from mlflow.projects import _resolve_experiment_id
 from mlflow.projects.utils import PROJECT_SYNCHRONOUS, PROJECT_DOCKER_ARGS, PROJECT_USE_CONDA
 from mlflow.store.tracking.file_store import FileStore
@@ -284,7 +285,7 @@ def test_parse_kubernetes_config():
     yaml_obj = None
     with open(kubernetes_config["kube-job-template-path"], 'r') as job_template:
         yaml_obj = yaml.safe_load(job_template.read())
-    kube_config = mlflow.projects._parse_kubernetes_config(kubernetes_config)
+    kube_config = _parse_kubernetes_config(kubernetes_config)
     assert kube_config["kube-context"] == kubernetes_config["kube-context"]
     assert kube_config["kube-job-template-path"] == kubernetes_config["kube-job-template-path"]
     assert kube_config["repository-uri"] == kubernetes_config["repository-uri"]
@@ -297,7 +298,7 @@ def test_parse_kubernetes_config_without_context():
         "kube-job-template-path": "kubernetes_job_template.yaml"
     }
     with pytest.raises(ExecutionException):
-        mlflow.projects._parse_kubernetes_config(kubernetes_config)
+        _parse_kubernetes_config(kubernetes_config)
 
 
 def test_parse_kubernetes_config_without_image_uri():
@@ -306,7 +307,7 @@ def test_parse_kubernetes_config_without_image_uri():
         "kube-job-template-path": "kubernetes_job_template.yaml"
     }
     with pytest.raises(ExecutionException):
-        mlflow.projects._parse_kubernetes_config(kubernetes_config)
+        _parse_kubernetes_config(kubernetes_config)
 
 
 def test_parse_kubernetes_config_invalid_template_job_file():
@@ -316,4 +317,4 @@ def test_parse_kubernetes_config_invalid_template_job_file():
         "kube-job-template-path": "file_not_found.yaml"
     }
     with pytest.raises(ExecutionException):
-        mlflow.projects._parse_kubernetes_config(kubernetes_config)
+        _parse_kubernetes_config(kubernetes_config)
